requirements.txt
matplotlib
numpy
spotipy
python-dotenv
matplotlib-venn
wordcloud
lyricsgenius
